+++ src/swap/defi/rango.ts
@@ -46,7 +46,7 @@ import {
   MakeTxParams,
   StringMap
 } from '../types'
-import { WEI_MULTIPLIER } from './defiUtils'
+import { createEvmApprovalEdgeTransactions, WEI_MULTIPLIER } from './defiUtils'
 
 const swapInfo: EdgeSwapInfo = {
   pluginId: 'rango',
@@ -570,7 +570,7 @@ export function makeRangoPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
 
     const providers = route.path.map(p => p.swapper.title)
 
-    let preTx: EdgeTransaction | undefined
+    const preTxs: EdgeTransaction[] = []
     let spendInfo: EdgeSpendInfo
 
     switch (tx.type) {
@@ -685,7 +685,7 @@ export function makeRangoPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
           fromNativeAmount: nativeAmount,
           metadataNotes,
           minReceiveAmount: route.outputAmountMin,
-          preTx,
+          preTxs,
           request,
           makeTxParams,
           swapInfo
@@ -699,32 +699,13 @@ export function makeRangoPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
         }
         const { approveData, approveTo } = evmTransaction
         if (approveData != null && approveTo != null) {
-          const approvalData = approveData.replace('0x', '')
-
-          spendInfo = {
-            tokenId: null,
-            memos: [{ type: 'hex', value: approvalData }],
-            spendTargets: [
-              {
-                nativeAmount: '0',
-                publicAddress: fromContractAddress
-              }
-            ],
-            assetAction: {
-              assetActionType: 'tokenApproval'
-            },
-            savedAction: {
-              actionType: 'tokenApproval',
-              tokenApproved: {
-                pluginId: fromWallet.currencyInfo.pluginId,
-                tokenId: fromTokenId,
-                nativeAmount
-              },
-              tokenContractAddress: fromContractAddress,
-              contractAddress: approveTo
-            }
-          }
-          preTx = await request.fromWallet.makeSpend(spendInfo)
+          const approvalTxs = await createEvmApprovalEdgeTransactions({
+            request,
+            approvalAmount: nativeAmount,
+            tokenContractAddress: fromContractAddress,
+            recipientAddress: approveTo
+          })
+          preTxs.push(...approvalTxs)
         }
         const customNetworkFee = {
           gasLimit:
@@ -791,7 +772,7 @@ export function makeRangoPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
       fromNativeAmount: nativeAmount,
       metadataNotes,
       minReceiveAmount: route.outputAmountMin,
-      preTx,
+      preTxs,
       request,
       spendInfo,
       swapInfo