+++ src/swap/defi/uni-v2-based/plugins/spookySwap.ts
@@ -137,7 +137,7 @@ export function makeSpookySwapPlugin(
     })
 
     let spendInfo = edgeSpendInfos[0]
-    let preTx: EdgeTransaction | undefined
+    const preTxs: EdgeTransaction[] = []
     if (edgeSpendInfos.length > 1) {
       spendInfo = edgeSpendInfos[1]
       const approvalSpendInfo: EdgeSpendInfo = {
@@ -157,7 +157,8 @@ export function makeSpookySwapPlugin(
         }
       }
 
-      preTx = await request.fromWallet.makeSpend(approvalSpendInfo)
+      const preTx = await request.fromWallet.makeSpend(approvalSpendInfo)
+      preTxs.push(preTx)
     }
 
     spendInfo = {
@@ -193,7 +194,7 @@ export function makeSpookySwapPlugin(
       swapInfo,
       fromNativeAmount: amountToSwap,
       expirationDate,
-      preTx
+      preTxs
     }
   }