+++ src/swap/defi/unizen.ts
@@ -31,7 +31,7 @@ import {
 import { convertRequest, getAddress } from '../../util/utils'
 import { EdgeSwapRequestPlugin, StringMap } from '../types'
 import { getTokenAddress } from './0x/util'
-import { getEvmApprovalData, WEI_MULTIPLIER } from './defiUtils'
+import { createEvmApprovalEdgeTransactions, WEI_MULTIPLIER } from './defiUtils'
 
 const pluginId = 'unizen'
 const swapInfo: EdgeSwapInfo = {
@@ -235,53 +235,28 @@ export function makeUnizenPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
     log('unizen spendInfo', JSON.stringify(spendInfo))
 
     // create approval transaction, if needed
-    let preTx: EdgeTransaction | undefined
+    const preTxs: EdgeTransaction[] = []
     const approvalNeeded = asMaybe(asInsufficientAllowance)(quote)
     if (approvalNeeded != null) {
-      const approvalData = await getEvmApprovalData({
-        contractAddress: spendParams.destinationAddress,
-        assetAddress: fromContractAddress,
-        nativeAmount: request.nativeAmount
-      })
-      if (approvalData == null)
-        throw new Error('Failed to create approval data')
-
-      const spendInfo: EdgeSpendInfo = {
-        tokenId: null,
-        memos: [{ type: 'hex', value: approvalData }],
-        spendTargets: [
-          {
-            nativeAmount: '0',
-            publicAddress: fromContractAddress
-          }
-        ],
+      const approvalTxs = await createEvmApprovalEdgeTransactions({
+        request,
+        approvalAmount: request.nativeAmount,
+        tokenContractAddress: fromContractAddress,
+        recipientAddress: spendParams.destinationAddress,
         networkFeeOption: spendParams.networkFeeOption,
-        customNetworkFee: spendParams.customNetworkFee,
-        assetAction: {
-          assetActionType: 'tokenApproval'
-        },
-        savedAction: {
-          actionType: 'tokenApproval',
-          tokenApproved: {
-            pluginId: request.fromWallet.currencyInfo.pluginId,
-            tokenId: request.fromTokenId,
-            nativeAmount: request.nativeAmount
-          },
-          tokenContractAddress: fromContractAddress,
-          contractAddress: spendParams.destinationAddress
-        }
-      }
-      preTx = await request.fromWallet.makeSpend(spendInfo)
+        customNetworkFee: spendParams.customNetworkFee
+      })
+      preTxs.push(...approvalTxs)
     }
 
-    log('unizen preTx', JSON.stringify(preTx ?? {}))
+    log('unizen preTxs', JSON.stringify(preTxs))
 
     const out: SwapOrder = {
       expirationDate: spendParams.expirationDate,
       fromNativeAmount: request.nativeAmount,
       metadataNotes: spendParams.metadataNotes,
       minReceiveAmount: spendParams.minReceiveAmount,
-      preTx,
+      preTxs,
       request,
       spendInfo,
       swapInfo